Adjust menu ui to use adjusted settings api
With the settings api now returning 404 instead of 200 on settings
that are not found, when requesting the currentStyle for
the user the request will now fail. This would be correct
however if the user does not have a default style. We therefore
have to set the default style on the 404 for the usersettings.